// dashboard/src/utils/hasura-api/generated/schemas/getEventInvocationLogsArgs.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface GetEventInvocationLogsArgs {
  /** Name of the event trigger */
  name: string;
  /** Name of the source database of the trigger */
  source?: string;
  /** Maximum number of invocation logs to be returned in one API call */
  limit?: number;
  offset?: number;
}