// dashboard/src/utils/hasura-api/generated/schemas/getEventInvocationLogsOperation.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { GetEventInvocationLogsArgs } from './getEventInvocationLogsArgs';
import type { GetEventInvocationLogsOperationType } from './getEventInvocationLogsOperationType';

/**
 * Operation used to fetch the invocation logs for a given event trigger.
 */
export interface GetEventInvocationLogsOperation {
  type: GetEventInvocationLogsOperationType;
  args: GetEventInvocationLogsArgs;
}